zephyr: rework non-secure TF-M related dependencies
Zephyr recently renamed various non-secure-related Kconfig symbols:

https://github.com/zephyrproject-rtos/zephyr/pull/36780

This breaks the memfault SDK defaults which are relying on the old
symbol names.

Resolve this by reworking the usages to just look at the SoC and
whether or not the non-secure image is being built, rather than keying
off of individual development boards provided by Nordic.

This introduces behavioral differences for other boards based off the
same SoC, but it seems more user friendly to have custom boards use
the same defaults as the dev boards, to ease porting and minimize
surprises.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>